Rename Project_2_VQE_Molecules/S5_Circuits-Copy1.ipynb to Project_2_VQE_Molecules/Research_Files/S5_Circuits-BK_IBMq-problem.ipynb
We run VQE simulations for H2 at 10 different bond_lengths and find the best parameters, then we send the circuit to the IBM q-device (ibmq_essex)
to record the overall computational time  and to validate the energy. To our surprise, the energy comes with large variation > 0.3 Ha and the time it takes for the job to be completed is often more than 10 minutes. [first observation of the variation]